NeuroBand_Final_Enhanced_Package/NeuroBand_Desktop_App/requirements.txt
tkinter
websocket-client==1.6.4
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
pyautogui==0.9.54
neurokit2==0.2.7
pygame==2.5.2
fpdf2==2.7.6
pyttsx3==2.90
Pillow==10.0.1
pandas==2.0.3
seaborn==0.12.2

